requirements.txt
Django>=4.0,<5.0
mysqlclient
djangorestframework

